requirements.txt
aiodns==1.1.1
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.2
Automat==20.2.0
beautifulsoup4==4.9.1
bs4==0.0.1
ccxt==1.30.62
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==2.9.2
dateparser==0.7.6
DateTime==4.3
html5lib==1.1
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
isort==4.3.21
mccabe==0.6.1
multidict==4.7.6
numpy==1.19.0
pandas==1.0.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybithumb==1.0.12
pycares==3.1.1
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
PyQt5==5.15.0
PyQt5-sip==12.8.0
pyqt5-tools==5.15.0.1.7
python-binance==0.7.5
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
pyupbit==0.2.6
regex==2020.6.8
requests==2.24.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
tzlocal==2.1
urllib3==1.25.9
webencodings==0.5.1
yarl==1.1.0
zope.interface==5.1.0
